Creacion de tipos genericos para la tabla reutilizable
