+++ src/facturacion/components/FacturacionTable.tsx
@@ -1,0 +1,3 @@
+export default function FacturacionTable() {
+    
+}
